refactoring: move readDirectory.js to the folderOperations folder
