+++ src/folderOperations/readDirectory.js
